Keep refactoring 06_tasks tests
